Add Web3AuthContext.tsx file to simplify auth.tsx and integrate in layout.tsx
